% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/iplots.R
\name{icorr}
\alias{icorr}
\title{icorr}
\usage{
icorr(mat, group, col, labels = NULL, cluster = TRUE,
  cor_method = "pearson", scatterplots = TRUE, plotOpts = NULL)
}
\arguments{
\item{mat}{data matrix (observations x variables)}

\item{group}{vector of grouping (\code{\link{factor}}-like) variables for
each observation}

\item{col}{a vector of colors for each unique \code{group} of points in
the scatter plots; note that colors will be assigned to the sorted levels
of \code{group}}

\item{labels}{optional character vector or named list of character vectors
to label each point; if \code{NULL}, points will be labeled by index}

\item{cluster}{logical or function; if \code{TRUE}, the variables will be
clustered and reordered; if \code{FALSE}, no reordering will be done;
otherwise, a custom clustering function may be given; see details}

\item{cor_method}{character string indicating which correlation coefficient
is to be computed; one of \code{'pearson'} (default), \code{'kendall'}, or
\code{'spearman'}: can be abbreviated; see \code{\link{cor}}}

\item{scatterplots}{logical; if \code{TRUE}, scatter plots of the linked
underlying data will be included}

\item{plotOpts}{list of additional plot options; see
\code{\link{icorrOpts}}}
}
\description{
Interactive correlation matrices with optional scatter plots.
}
\details{
If \code{col} is given with no \code{group} variable, the colors for each
observation will be recycled in order.

Default cluster method is \code{stats::hclust(dist(x), method = 'average')}
which will return a list containing a named vector, \code{"order"}, which is
used to reorder the variables.

In order to pass a custom clustering function to \code{cluster}, the
function must take a single input (a correlation matrix) and return either
a vector or a list with a named vector, \code{"order"}.
}
\examples{
## heatmap only
icorr(mtcars, scatterplots = FALSE)

icorr(mtcars, group = mtcars$cyl, col = c('blue','red','green'),
      plotOpts = list(corcolors = heat.colors(3)))

## larger matrix example
set.seed(1)
dat <- replicate(50, mtcars[, sample(1:11, 1), drop = FALSE])
dat <- do.call('cbind', dat)
icorr(dat, cluster = TRUE, group = mtcars$cyl)

## using a custom clustering function
icorr(dat, cluster = function(x) hclust(dist(x, method = 'maximum')))
icorr(dat, cluster = function(x) sample(seq.int(ncol(x))))
icorr(dat, cluster = function(x) list(order = order(rowMeans(x))))
}
\seealso{
\code{\link{iscatter}}, \code{\link{idot}}, \code{\link{itree}},
\code{\link{icurve}}, \code{\link[qtlcharts]{iplotCorr}}
}

