% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{icorrOpts}
\alias{icorrOpts}
\title{icorr options}
\description{
Below is a list of options for correlation matrices with optional
linked scatter plots (\code{\link{icorr}}).
}
\details{
\tabular{rlll}{
\tab \strong{option} \tab \strong{description} \tab \strong{default} \cr
\tab \code{height} \tab height of each panel in pixels  \tab \code{560} \cr
\tab \code{width} \tab total width of panels \tab \code{1050} \cr
\tab \code{margin} \tab margins in pixels (left, top, right, bottom, inner)
\tab \code{list(left=70, top=40, right=5, bottom=70, inner=5)} \cr
\tab \code{corcolors} \tab heat map colors (same length as \code{zlim}) \tab
\code{c('darkslateblue', 'white', 'crimson')} \cr
\tab \code{zlim} \tab z-axis limits \tab \code{c(-1, 0, 1)}\cr
\tab \code{rectcolor} \tab color of background rectangle \tab
\code{'#E6E6E6'} \cr
\tab \code{cortitle} \tab title for heat map panel \tab \code{''} \cr
\tab \code{scattitle} \tab title for scatterplot panel \tab \code{''} \cr
\tab \code{scatcolors} \tab vector of point colors for scatterplot \tab 
\code{NULL} \cr
}
}
