% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iplots.R
\name{icurve}
\alias{icurve}
\title{icurve}
\usage{
icurve(mat, labels, group, iscatter1 = NULL, iscatter2 = NULL,
  plotOpts = NULL, digits = NULL)
}
\arguments{
\item{mat}{a numeric matrix, \code{observations x timepoints}}

\item{labels}{optional character vector or named list of character vectors
to label each point; if \code{NULL}, points will be labeled by index}

\item{group}{vector of group (\code{\link{factor}}-like) variables}

\item{iscatter1}{a numeric matrix or \code{NULL}, \code{nobs x 2}}

\item{iscatter2}{see \code{iscatter1}}

\item{plotOpts}{list of additional plot options; see
\code{\link{icurveOpts}}}

\item{digits}{integer indicating number of significant digits to use}
}
\description{
Interactive curves over time with optional interactive scatter plots.

Each timepoint is a column of \code{mat}, that is, times are treated as
discrete and assumed to be uniformly-spaced. Observations will be linked
across all plots (curves and optional scatter plots) by row index, so 
observations should be in the same order in all data provided.
}
\examples{
## basic usage, no scatter plots
icurve(matrix(AirPassengers, ncol = 12, byrow = TRUE), labels = 1949:1960)

## connecting observations over time with additional linked scatter plots
set.seed(1)
n <- 25
## 25 observations and 5 timepoints
mm <- cbind(runif(n, 0, 10), runif(n, 10, 50), runif(n, 30, 60),
            runif(n, 40, 80), runif(n, 60, 100))
mm <- mm[order(mm[, 5]), ]

## two sets of data (25 obs x 2 columns: id and numeric data)
x1 <- cbind(1:n, kinda_sort(rnorm(n), n / 2))
x2 <- cbind(1:n, kinda_sort(sample(1:100, n), n / 2))

icurve(mat = mm, group = 1:25 \%\% 5 == 0,
       iscatter1 = x1, iscatter2 = x2,
       labels = list(
         Patient = 1:25,
         Disease = sample(c('ALL','CLL'), 25, replace = TRUE)),
       plotOpts = list(
         curves_xlab = 'Response evaluation time point',
         curves_ylab = '\% response',
         scat1_xlab = 'Patient', scat1_ylab = 'Lab var 1',
         scat2_xlab = 'Patient', scat2_ylab = 'Lab var 2'))

}
\seealso{
\code{\link{icorr}}, \code{\link{idot}}, \code{\link{itree}},
\code{\link{icurve}}, \code{\link[qtlcharts]{iplotCurves}}
}

