% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{icurveOpts}
\alias{icurveOpts}
\title{icurve options}
\description{
Below is a list of options for tree plots (\code{\link{icurve}}).
}
\details{
\tabular{rlll}{
\tab \strong{option} \tab \strong{description} \tab \strong{default} \cr

\tab \code{width} \tab width of plot in pixels \tab \code{1000} \cr
\tab \code{height} \tab height of plot in pixels \tab \code{600} \cr
\tab \code{margin} \tab margins in pixels (left, top, right, bottom,
inner) \tab \code{list(left=60, top=40, right=100, bottom=40, inner=10)} \cr
\tab \code{axispos} \tab position of axis labels in pixels (xtitle, ytitle,
xlabel, ylabel) \tab \code{list(xtitle=25, ytitle=30, xlabel=5, ylabel=5)} \cr
\tab \code{titlepos} \tab position of plot title in pixels \tab \code{20} \cr
\tab \code{ylim} \tab y-axis limits \tab \code{NULL} \cr
\tab \code{nyticks} \tab no. ticks on y-axis \tab \code{5} \cr
\tab \code{yticks} \tab vector of tick positions on y-axis \tab
\code{NULL} \cr
\tab \code{tickwidth} \tab width of tick marks at markers, in pixels
\tab \code{10} \cr
\tab \code{rectcolor} \tab color of background rectangle \tab
\code{'#E6E6E6'} \cr
\tab \code{linecolor} \tab color of lines \tab \code{'slateblue'} \cr
\tab \code{linecolorhilit} \tab color of lines when highlighted \tab
\code{'Orchid'} \cr
\tab \code{linewidth} \tab width of lines \tab \code{3} \cr
\tab \code{title} \tab title for plot \tab \code{''} \cr
\tab \code{xlab} \tab x-axis label \tab \code{''} \cr
\tab \code{ylab} \tab y-axis label \tab \code{''} \cr
}
}

