% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iplots.R
\name{idot}
\alias{idot}
\title{idot}
\usage{
idot(x, y = NULL, group = NULL, cex = 3, xlim = NULL, ylim = NULL,
  xlab = NULL, ylab = NULL, main = NULL, labels = NULL,
  plotOpts = NULL, digits = NULL)
}
\arguments{
\item{x}{a numeric vector of y values (if \code{y} is \code{NULL}) or if
\code{y} is given, a vector of x values; if \code{x} is an named vector,
these labels will be passed on to \code{labels}}

\item{y}{optional numeric vector of y values; if missing, \code{x} is used}

\item{group}{optional vector of categories for coloring points}

\item{cex}{point size in pixels}

\item{xlim, ylim}{x- and y-limits}

\item{xlab, ylab, main}{the x-, y-, and main labels}

\item{labels}{optional character vector or named list of character vectors
to label each point; if \code{NULL}, points will be labeled by index}

\item{plotOpts}{list of additional plot options; see
\code{\link{idotOpts}}}

\item{digits}{integer indicating number of significant digits to use}
}
\description{
Interactive dot plot.

Any numeric data can be plotted on the y-axis but the x-axis should be
discrete groups. Points will be labeled with the vector names, the
\code{labels} parameter, or by index otherwise.
}
\examples{
## basic usage with named vectors
idot(setNames(mtcars$mpg, rownames(mtcars)))

## using group and subgroup (a vector of logicals to color points)
with(mtcars,
     idot(gear, mpg, group = am + 1L, cex = 5, ylim = c(0, 40),
          labels = rownames(mtcars)))

## labels parameter supports additional formatting with html tags
with(ivolcano,
     idot(substr(rownames(ivolcano), 1, 1), logFC,
          group = cut(pval, c(0, 0.05, .1, 1)), xlab = 'Treatment arm',
          labels = list(
            ' ' = rownames(ivolcano),
            'log2(FC)' = round(logFC, 2),
            '<i>p</i>-value' = format.pval(ivolcano$pval, digits = 2,
                                           eps = .05))))

}
\seealso{
\code{\link{iscatter}}, \code{\link{icorr}}, \code{\link{itree}},
\code{\link{icurve}}, \code{\link[qtlcharts]{iplotPXG}}
}

