% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iplotr-package.R
\docType{package}
\name{iplotr-package}
\alias{iplotr-package}
\alias{iplotr}
\title{Interactive plotting functions}
\description{
Generic interactive plots (using D3.js) based on \pkg{\link{qtlcharts}}.
}
