% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iplots.R
\name{iscatter}
\alias{iscatter}
\title{iscatter}
\usage{
iscatter(x, y = NULL, group, col, cex = 3, xlim = NULL, ylim = NULL,
  xlab = NULL, ylab = NULL, main = NULL, labels = NULL,
  plotOpts = NULL, digits = NULL)
}
\arguments{
\item{x, y}{vectors of x- and y-coordinates; if \code{y} is \code{NULL},
\code{x} will be plotted along the y-axis by index of \code{x}}

\item{group}{vector of group (\code{\link{factor}}-like) variables used for
additional aesthetics such as \code{col}}

\item{col}{a vector of colors for each unique \code{group}; note that colors
will be assigned to the sorted levels of \code{group}}

\item{cex}{point size in pixels}

\item{xlim, ylim}{x- and y-limits}

\item{xlab, ylab, main}{the x-, y-, and main labels}

\item{labels}{optional character vector or named list of character vectors
to label each point; if \code{NULL}, points will be labeled by index}

\item{plotOpts}{list of additional plot options; see
\code{\link{iscatterOpts}}}

\item{digits}{integer indicating number of significant digits to use}
}
\description{
Interactive scatter plot.

If \code{labels} is a vector, points are labeled as-is and recycled if
necessary. Optionally, \code{labels} can be a \emph{named} list where each
element of the list is a character vector (or one which can be coerced)
having the same length of \code{x}; see examples. If \code{x} and \code{y}
are named vectors, points will be labeled without needing to use
\code{labels}.
}
\examples{
## basic usage
iscatter(1:5, col = 1:5, cex = 5)

## using named vectors
x <- with(mtcars, setNames(mpg, mpg))
y <- with(mtcars, setNames(wt, wt))
iscatter(x, y)

## using the labels parameter
with(mtcars,
  iscatter(wt, mpg, group = cyl, col = 1:3, cex = 5,
           main = 'Motor Trend car road tests',
           labels = list(model = rownames(mtcars), mpg = mpg, hp = hp)))

## compare with base graphics
with(mtcars, plot(wt, mpg, col = factor(cyl), pch = 19))

## labels parameter supports additional formatting with html tags
with(ivolcano,
     iscatter(logFC, -log10(pval), group = pval < 0.05 & abs(logFC) > 1,
              col = c('lightgrey','green'),
              labels = list(
                ' ' = rownames(ivolcano),
                'log<sub>2</sub>(FC)' = round(logFC, 2),
                '<i>p</i>-value' = format.pval(ivolcano$pval, digits = 2,
                                               eps = .05))))

}
\seealso{
\code{\link{icorr}}, \code{\link{idot}}, \code{\link{itree}},
\code{\link{icurve}}, \code{\link[qtlcharts]{iplot}}
}

