% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{iscatterOpts}
\alias{iscatterOpts}
\title{iscatter options}
\description{
Below is a list of options for scatter plots (\code{\link{iscatter}}).
}
\details{
\tabular{rlll}{
\tab \strong{option} \tab \strong{description} \tab \strong{default} \cr
\tab \code{height} \tab height of plot in pixels \tab \code{500} \cr
\tab \code{width} \tab width of plot in pixels \tab \code{800} \cr
\tab \code{title} \tab title for plot \tab \code{''} \cr
\tab \code{margin} \tab margins in pixels (left, top, right, bottom, inner)
\tab \code{list(left=60, top=40, right=40, bottom=40, inner=5)} \cr
\tab \code{xlab} \tab x-axis label \tab \code{'X'} \cr
\tab \code{ylab} \tab y-axis label \tab \code{'Y'} \cr
\tab \code{axispos} \tab position of axis labels in pixels (xtitle, ytitle,
xlabel, ylabel) \tab \code{list(xtitle=25, ytitle=30, xlabel=5, ylabel=5)} \cr
\tab \code{titlepos} \tab position of plot title in pixels \tab \code{20} \cr
\tab \code{xlim} \tab x-axis limits \tab \code{NULL} \cr
\tab \code{xticks} \tab vector of tick positions on x-axis \tab
\code{NULL} \cr
\tab \code{nxticks} \tab number ticks on x-axis \tab \code{5} \cr
\tab \code{ylim} \tab y-axis limits \tab \code{NULL} \cr
\tab \code{yticks} \tab vector of tick positions on y-axis \tab
\code{NULL} \cr
\tab \code{nyticks} \tab number ticks on y-axis \tab \code{5} \cr
\tab \code{rectcolor} \tab color of background rectangle \tab
\code{'#E6E6E6'} \cr
\tab \code{pointcolor} \tab colors for points \tab \code{NULL} \cr
\tab \code{pointsize}\tab size of points in pixels \tab \code{3} \cr
\tab \code{pointstroke} \tab color of outer circle for points \tab
\code{'black'} \cr
\tab \code{rotate_ylab} \tab whether to rotate the y-axis label \tab
\code{NULL} \cr
\tab \code{xNA} \tab treatment of missing values (handle=T/F, force=T/F,
width, gap) \tab \code{list(handle=true, force=false, width=15, gap=10)} \cr
\tab \code{yNA} \tab treatment of missing values (handle=T/F, force=T/F,
width, gap) \tab \code{list(handle=true, force=false, width=15, gap=10)} \cr
}
}
