% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/iplots.R
\name{itree}
\alias{itree}
\title{itree}
\usage{
itree(y, group, ylim = NULL, xlab = NULL, ylab = NULL, main = NULL,
  plotOpts = NULL)
}
\arguments{
\item{y}{a \emph{uniquely-named} numeric vector}

\item{group}{vector of grouping (\code{\link{factor}}-like) variables the
same length as \code{y}}

\item{ylim}{y-axis limits}

\item{xlab,ylab,main}{the x-, y-, and main labels}

\item{plotOpts}{list of additional plot options; see
\code{\link{itreeOpts}}}
}
\description{
Interactive tree plot with search bar.
}
\examples{
itree(setNames(rnorm(10), letters[1:10]))
itree(setNames(rnorm(20), letters[1:20]), 1:4)

set.seed(1)
n <- 100
ng <- 5
gr <- sort(sample(LETTERS[1:ng], n, replace = TRUE))
yv <- kinda_sort(runif(n, -1, 1), n = n / 2) * 100
itree(yv, gr, main = 'Subject response from baseline',
      ylim = c(-100,100), xlab = 'Treatment arm', ylab = '\% change')
}
\seealso{
\code{\link{iscatter}}, \code{\link{icorr}}, \code{\link{itree}},
\code{\link{icurve}}, \code{\link[qtlcharts]{iplotMap}}
}

