% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/iplotr-package.R
\docType{data}
\name{ivolcano}
\alias{ivolcano}
\title{ivolcano data}
\format{A data frame with 500 rows and 2 variables:

\tabular{lll}{
\tab \code{logFC} \tab \code{\link{log2}} fold-change \cr
\tab \code{pval} \tab p-values corresponding to the fold-change \cr
}}
\usage{
ivolcano
}
\description{
Generated data for creating interactive volcano plots in \code{\link{idot}}
and \code{\link{iscatter}} examples.
}
\keyword{datasets}

