% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/iplotr-package.R
\name{plotOpts}
\alias{iplotOpts}
\alias{plotOpts}
\title{Plot options}
\description{
In additional to the arguments passed directly to the individual plotting
functions, certain aesthetics of each can be changed by using the
\code{plotOpts} argument. This should be a \emph{list} of options such as
\code{height} or \code{width} setting the plot size in pixels;
\code{rectcolor} for the background color of the plot; \code{margin} to
specify the left, top, right, bottom, and inner margins; or \code{axispos}
and \code{titlepos} to specify the label positions in pixels.
}
\details{
Each type of plot has its own additional aesthetic options:

\tabular{rll}{
\tab \code{\link{iscatter}} \tab scatter plot (\code{\link{iscatterOpts}}) \cr
\tab \code{\link{icorr}} \tab correlation matrix and optional
scatterplots (\code{\link{icorrOpts}}) \cr
\tab \code{\link{idot}} \tab dot plots (\code{\link{idotOpts}}) \cr
\tab \code{\link{itree}} \tab tree plots (\code{\link{itreeOpts}}) \cr
\tab \code{\link{icurve}} \tab curves with optional scatterplots
(\code{\link{icurveOpts}}) \cr
}

See \url{http://kbroman.org/qtlcharts/assets/vignettes/chartOpts.html} or
the help pages for more plotting options.
}

