% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iplots.R
\name{idot}
\alias{idot}
\title{idot}
\usage{
idot(y, group, cex = 3, subgroup, xlim = NULL, ylim = NULL, xlab = NULL,
  ylab = NULL, main = NULL, labels = NULL, plotOpts = NULL)
}
\arguments{
\item{y}{numeric vector}

\item{group}{vector of grouping (\code{\link{factor}}-like) variables the
same length as \code{y}}

\item{cex}{point size in pixels}

\item{subgroup}{a logical vector the same length as \code{y}}

\item{xlim, ylim}{x- and y-limits}

\item{xlab, ylab, main}{the x-, y-, and main labels}

\item{labels}{optional character vector or named list of character vectors
to label each point; if \code{NULL}, points will be labeled by index}

\item{plotOpts}{list of additional plot options; see
\code{\link{idotOpts}}}
}
\description{
Interactive dot plot.

Any numeric data can be plotted on the y-axis but the x-axis should be
discrete groups. Points will be labeled with the vector names, the
\code{labels} parameter, or by index otherwise.
}
\examples{
## basic usage with named vectors
idot(setNames(mtcars$mpg, rownames(mtcars)))

## using group and subgroup (a vector of logicals to color points)
with(mtcars,
     idot(mpg, gear, subgroup = am == 1, cex = 5, ylim = c(0, 40),
          labels = rownames(mtcars)))

## labels parameter supports additional formatting with html tags
with(ivolcano,
     idot(logFC, substr(rownames(ivolcano), 1, 1),
          subgroup = pval < 0.05, xlab = 'Treatment arm',
          labels = list(
            ' ' = rownames(ivolcano),
            'log2(FC)' = round(logFC, 2),
            '<i>p</i>-value' = format.pval(ivolcano$pval, digits = 2,
                                           eps = .05))))

}
\seealso{
\code{\link{iscatter}}, \code{\link{icorr}}, \code{\link{itree}},
\code{\link{icurve}}, \code{\link[qtlcharts]{iplotPXG}}
}

