% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iplots.R
\name{itree}
\alias{itree}
\title{itree}
\usage{
itree(
  y,
  group,
  ylim = NULL,
  xlab = NULL,
  ylab = NULL,
  main = NULL,
  plotOpts = NULL,
  digits = NULL
)
}
\arguments{
\item{y}{a \emph{uniquely-named} numeric vector}

\item{group}{vector of grouping (\code{\link{factor}}-like) variables the
same length as \code{y}}

\item{ylim}{y-axis limits}

\item{xlab, ylab, main}{the x-, y-, and main labels}

\item{plotOpts}{list of additional plot options; see
\code{\link{itreeOpts}}}

\item{digits}{integer indicating number of significant digits to use}
}
\description{
Interactive tree plot with search bar.

Any numeric data can be plotted on the y-axis but the x-axis should be
discrete groups. Points will be labeled with their \emph{unique} names
and y-value.
}
\examples{
## basic usage with and without groups
itree(setNames(rnorm(10), letters[1:10]))
itree(setNames(rnorm(20), letters[1:20]), group = 1:4)

## another example
set.seed(1)
n <- 100
ng <- 5
gr <- sort(sample(LETTERS[1:ng], n, replace = TRUE))
yv <- kinda_sort(runif(n, -1, 1), n = n / 2) * 100
itree(yv, gr, main = 'Subject response from baseline',
      ylim = c(-100,100), xlab = 'Treatment arm', ylab = '\% change')

## search is more useful with lots of data
set.seed(1)
p0 <- function(x) paste0(x, collapse = '')
n <- 500
gr <- replicate(n, (function()
  p0(sample(c('A','T','C','G'), 15, replace = TRUE)))())
y <- setNames(rpois(n, 40), gr)

## hacky way to remove tick labels which start running together
gr_labels <- Map(function(x) p0(rep(' ', x)), sequence(n))

itree(y, unlist(gr_labels), ylab = 'Frequency', xlab = 'Sequence',
      plotOpts = list(tickwidth = 2, linewidth = 5, width = 2000))

}
\seealso{
\code{\link[qtlcharts]{iplotMap}} from the \pkg{qtlcharts} package

\code{\link[htmlwidgets]{saveWidget}}

Other iplots: 
\code{\link{icorr}()},
\code{\link{icurve}()},
\code{\link{idot}()},
\code{\link{iscatter}()}
}
\concept{iplots}
